pyproject.toml
[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "video_archive"
description = "Video Archive"
version = "0.1.0"
readme = "README.md"
authors = []


[tool.poetry.dependencies]
python = "3.8.5"

# TODO: remove after https://github.com/python-poetry/poetry/issues/2042
importlib-metadata = "<1.2.0,>=1.1.3"

django = "^3.0.9"
django-split-settings = "^1.0"
django-axes = "^5.4"
django-csp = "^3.5"
django-health-check = "^3.11"
django-http-referrer-policy = "^1.0"
django-feature-policy = "^3.4"

psycopg2 = "^2.8"
gunicorn = "^20.0"
python-decouple = "^3.3"
bcrypt = "^3.1"
structlog = "^20.1.0"
typing_extensions = "^3.7.4"
celery = "^4.4.7"
flower = "^0.9.5"
djangorestframework = "^3.11.1"
isort = "^5.3.2"
add-trailing-comma = "^2.0.1"
preview-generator = "^0.14"
ffprobe = "^0.5"
PythonVideoConverter = "^1.0.3"
uvicorn = "^0.11.8"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^2.1"
django-querycount = "^0.7"
django-migration-linter = "^2.4"
django-coverage-plugin = "^1.8"
nplusone = "^1.0"

wemake-python-styleguide = "^0.14"
flake8-pytest-style = "^1.2"
flake8-django = "^1.1"
flake8-logging-format = "^0.6"
nitpick = "^0.22"

pytest = "^6.0"
pytest-django = "^3.9"
pytest-cov = "^2.10"
pytest-randomly = "^3.4"
pytest-deadfixtures = "^2.2"
pytest-testmon = "^1.0"
pytest-timeout = "^1.4"
django-test-migrations = "^1.0"
hypothesis = "^5.23"

mypy = "^0.770"
django-stubs = "^1.4"

ipython = "^7.17"

sphinx = "^2.4"
sphinx-autodoc-typehints = "^1.10"
doc8 = "^0.8"

yamllint = "^1.24"
safety = "^1.9"
dotenv-linter = "^0.1"
polint = "^0.4"
dennis = "^0.9"
dump-env = "^1.2"
